Add support for completing workflow tasks
A new feature has been introduced to handle completion of workflow tasks. This included the addition of a new 'complete' method in the workflows task service, the WorkflowTaskAlreadyCompletedException for handling task already completed errors, and the WorkflowTaskCompleteResult for managing the task completion result. The CHANGELOG and documentation have been updated accordingly.

Signed-off-by: mesilov <mesilov.maxim@gmail.com>