Remove destructor in Response.php
The destructor was removed from the Response.php file because it included logging that was deemed unnecessary. The associated logging for responseInfo, networkTimings, and restTimings was relying on possibly uninvoked getResponseData method and an unfinished asynchronous request. This operation makes the logic of the code more straightforward and avoid redundancy.

Signed-off-by: mesilov <mesilov.maxim@gmail.com>