Refactor variable and function names in telephony tests
The commit primarily addresses the renaming of variables and adjusting the visibility of setUp functions in the Telephony test classes. Variables such as '$sb' and '$cost' have been renamed to more meaningful names like '$serviceBuilder' and '$money'. The visibility of 'setUp' functions has been changed from public to protected for better encapsulation.

Signed-off-by: mesilov <mesilov.maxim@gmail.com>