Add delete method in ExternalLine service
A delete method has been implemented in the ExternalLine service, which allows to delete an external line. The method is tested in the ExternalLineTest class where a line is added, verified, deleted, and the deletion is confirmed. A new result object, EmptyResult, has been introduced to cover cases where a function doesn't return any specific data.

Signed-off-by: mesilov <mesilov.maxim@gmail.com>