Add default arguments handling in User.get method
A check has been added in User.get method, to set a default argument as ascending order of IDs when no explicit order is provided. A corresponding test case 'test get users list with default arguments' has also been included to validate the functionality. The CHAGELOG.md is updated to reflect this fix.

Signed-off-by: mesilov <mesilov.maxim@gmail.com>