Replace `ramsey/uuid` with `symfony/uid` package
The existing dependency on `ramsey/uuid` has been replaced with `symfony/uid` in the codebase. This change is reflected in the `Command` constructor of `src/Core/Commands/Command.php`, and anywhere else where the ID of a command was previously retrieved using `getUuid()`. The `getName()` method has also been replaced with `getId()`. The composer.json and CHANGELOG files have been updated correspondingly.

Signed-off-by: mesilov <mesilov.maxim@gmail.com>