Refactor AccessToken to AuthToken in SDK tests
The AccessToken and related classes and interfaces have been replaced with AuthToken and relevant entities across the SDK tests. The main aim was to unify the nomenclature and make it more expressive and intuitive. For instance, the AccessTokenFileStorage has been renamed to AuthTokenFileStorage, and methods and variables have been renamed accordingly.

Signed-off-by: mesilov <mesilov.maxim@gmail.com>