Add application installation interface and Bitrix24 ID to contact person
This update adds an application installation interface for setting up interactions between the application and the Bitrix24 API. A Bitrix24 user ID field is also added to the contact person interface. This modification extends the functionality of the contact person, enabling more advanced data tracking. The new ID field maps a Bitrix24 user with a contact person, if there is any association.

Signed-off-by: mesilov <mesilov.maxim@gmail.com>